Add SMS attributes to instance
The XML output now includes the SMS attributes as they are used by
JavaRosa in Medic Collect.

The `sms_keyword` and `sms_separator` are optional settings (according
to https://formhub.org/syntax/) and not brought through to the xform
xml. In JavaRosa these are represented as `prefix` and `delimiter`
attributes for the children of <instance>, which is why the attributes
were names as such here.